1.) Added default mimetype value for missing values 2.) Fixed the port type on the serverwelcome() to support 32-bit architectures

+++ src/core.jl
@@ -29,7 +29,7 @@ oxygen_title = raw"""
 
 """
 
-function serverwelcome(host::String, port::Int64)
+function serverwelcome(host::String, port::Int)
     printstyled(oxygen_title, color = :blue, bold = true)  
     @info "✅ Started server: http://$host:$port" 
     @info "📖 Documentation: http://$host:$port$docspath"
